Glue_with_step/requirements.txt
pandas
numpy
sqlalchemy
dotenv
pymysql
mysql-connector-python
dlt[filesystem]
dlt
s3fs
boto3
pyarrow